+++ CHANGELOG.md
@@ -14,7 +14,7 @@
     * Added `DATECREATED` and `DATEMODIFIED` user fields
     * Added `DATECREATED` and `DATEMODIFIED` website fields
     * Added handling of the time of the last website update
-    * Added **Page Component** feature
+    * Added **Page Component** feature. Disabled by default. Need to uncomment `# define('GSPAGECOMPONENT', true);` in *gsconfig.php* to enable it
     * Added `get_page_component()` theme function to eval page component code stored in `component` field of the current page
     * Added `getPageComponent()` cache function to eval page component code of the requested page
     * Added `returnPageComponent()` cache function to return component code of the requested page in plain text